Agregue lo de singleton (ahora si)

+++ Ingee/src/main/java/Clases/Empleado.java
@@ -7,7 +7,7 @@ import java.util.List;
 public class Empleado {
     private int id;
     private String nombre;
-    private int telefono;
+    private long telefono;
     private float sueldo;
     private List<Trabajo> trabajos = new ArrayList<>();
     private List<Retiro> retiros = new ArrayList<>();
@@ -15,7 +15,7 @@ public class Empleado {
     public Empleado() {}
 
     // Constructor completo
-    public Empleado(int id, String nombre, int telefono, float sueldo, List<Trabajo> trabajos, List<Retiro> retiros) {
+    public Empleado(int id, String nombre, long telefono, float sueldo, List<Trabajo> trabajos, List<Retiro> retiros) {
         this.id = id;
         this.nombre = nombre;
         this.telefono = telefono;
@@ -31,8 +31,8 @@ public class Empleado {
     public String getNombreEmpleado() { return nombre; }
     public void setNombreEmpleado(String nombre) { this.nombre = nombre; }
 
-    public int getTelefonoEmpleado() { return telefono; }
-    public void setTelefonoEmpleado(int telefono) { this.telefono = telefono; }
+    public long getTelefonoEmpleado() { return telefono; }
+    public void setTelefonoEmpleado(long telefono) { this.telefono = telefono; }
 
     public float getSueldoEmpleado() { return sueldo; }
     public void setSueldoEmpleado(float sueldo) { this.sueldo = sueldo; }
